Split KotlinGenTask file
